Refactor a lot of code in the terminal window.
Remove the tab-label object and move the code into 
terminal-screen. This avoids a bunch of complexity 
in signals and bindings when adding and removing tabs.

Use the UI manager for the right-click tab- and goto-menu, 
making the code easier to understand. We now completely 
use GtkActions and the UI manager.

Do not unset focusing in the notebook, but move the focus 
back to the terminal on button-release events. This allows 
enabling of notebook scrolling.

Use a pointer to the active terminal screen, this avoids 
a bunch of calls to the active tab lookup function.

Change the handling of the geometry size updates. Hopefully 
this will fix the issue where the terminal was getting 
smaller and smaller when showing/hiding the tabbar (bug #4728).

Make tab dnd code more robust and fix some crashes.

Various function renames and some extra debugging in screen 
functions which had bug reports with segfaults in launchpad, 
space indents etc.


(Old svn revision: 30134)