Catch keybindings from previous and next tab when cycling is disabled (bug #3715).
Maybe not the best solution, therefore only activated for the next
and previous tab actions. The idea is that if we receive an accel-activate
event from the accel group after gtk has checked the group, we
check if it is the same accelerator of the insensitive {next,prev}-tab, if
so we return TRUE, so no odd characters are printed in vte with
the default <Ctrl>+{Pageup,Pagedown} shortcuts.