Initial man page for terminal.
Instead of providing Terminal --help we install a complete
man-page for help. The docbook from which the man is generated
is also used in the html documentation.

This makes it easier to translate (no space indentation like
we have in --help atm) and sharing the command line options
help with the html pages.