Avoid racing on the size-changed signal (bug #6921).
Some internals in vte changed, which causes size-allocate signals on background
changes, resulting in the terminal allocating and leaking (because there is still
a leak in vte) backgrounds.
Fix this my doing the background update in an idle, which seems to resolve the
problem and to be sure, don't watch size changes when a transparent or plain
background is used.