Refine the MiscMiddleClickOpensUri option again: set if off by default
It seems that opening URIs by middle click brings a lot of inconvenience to
users as the same button is used for selection paste. Besides, other terminal
emulators, such as gnome-terminal or qterminal, appear not to be using middle
click for this action.
This change makes Ctrl + left click the default action for opening URIs,
providing users with a possibility to change it to middle click via a hidden
preference.