Fix background color changing when changing focus
If "Vary the background color for each tab" is enabled.

This was happening due to the "style-updated" signal being emitted on each
focus in/out event, as well as the actual theme change.
Since there's no way to tell whether or not the theme has been changed, the
solution is to remember that a tab already has a random background color and
prevent it from generating a new one.

Fixes bug #15740