wayland: Fix new window size
XParseGeometry() can be run on Wayland without any problems at present,
and probably in the future, since it's essentially low-level parsing
that's unlikely to change and there's nothing really X11-specific about
it in reality.

That said, running X11 code on Wayland remains problematic, hence the
restriction of its use to X11. On the other hand,
gtk_window_parse_geometry(), which is used as a replacement on Wayland,
does essentially the same thing (with a copy of the X11 code in GTK that
we're sure won't change to make it impossible to run on Wayland).

This does require some adaptation, however, as width and height remain
internal to gtk_window_parse_geometry().