wayland: Fix new window size
The result of gtk_window_parse_geometry(), in conjunction with
gtk_window_set_geometry_hints(), depends on the compositor used on
Wayland and is therefore unreliable. On the other hand, the
XParseGeometry() code is not really X11-specific and should not be
subject to many changes in the future. So we're using a copy of this
code made in GTK 3.24.38 to extend XParseGeometry() to Wayland.