Do not reconnect accels when closing last tab
G_IS_OBJECT and similar derived macros are not valid tests of whether an
object is still alive. They are intended to test the type of an object
assumed to be alive (although they return FALSE for a null pointer for
convenience).

This can cause a crash on some systems when closing last tab if
/misc-always-show-tabs is TRUE.

Fixes: 31712a009ec4db4a0731431905a169fe610286e5