Select text in Find dialog when it is shown
I will often have searched for something in a terminal window at some
point in the past, and then some time later (perhaps even days later),
hit ctrl+shift+f to start a new search, and immediately start typing my
new search query, only to find that the previous text in the dialog
entry is still there, and I've merely appended to it.  Then I have to
ctrl+a to select all and retype it.  Annoying.

This changes the behavior so that every time the search dialog is shown,
if there is not already a selection in the entry, it selects the text so
that immediate typing will overwrite it.