window: Add dnd support outside screen
This allows you to drop a tab anywhere on the window, so that it fits in
the expected place in the notebook, and not just when it's dropped on
the screen. It also avoids unwanted tab detachment when a dnd is
unintentionally initiated to the menubar or toolbar by simply clicking
on a tab to activate it.