Allow handling middle click event by terminal applications
Currently Xfce terminal always handles the middle button click event
on its own (pastes the selection) instead of passing it to the terminal
application even if mouse tracking is enabled by the application, so
applications (e.g. vim) are unable to handle middle clicks. So let it
pass this event to the application like other mouse click events.

To achieve that, just pass it to the VteTerminal's event handler which
will return true if mouse tracking is enabled and thus the event was
passed to the application.

Note that the VteTerminal's handler also pastes the selection on its own
(as long as the GTK setting gtk-enable-primary-paste is enabled) and
in such case it also returns true. So we should only paste the selection
if it returns false, to avoid duplicated paste.